% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP_func.R
\name{DGP}
\alias{DGP}
\title{Data Generate Process}
\usage{
DGP(n, p, example, case)
}
\arguments{
\item{n}{sample size}

\item{p}{dimension}

\item{example}{example: 1, 2, 3}

\item{case}{case: "a", "b", "c"}
}
\value{
the data \eqn{X}.
}
\description{
Data Generate Process
}
\examples{
n = 50; p = 10; example = 1; case = "a"
DGP(n, p, example, case)
}
